+++ docs/changelog.md
@@ -5,6 +5,12 @@
 ## Changelog
 
 
+### v0.4.0 alpha release
+
+Support yWriter items as NovelWriter objects.
+
+Based on PyWriter v3.32.0.
+
 ### v0.2.0 First public alpha release
 
 Based on PyWriter v3.32.0.